copy updates for new toggles and clean up admin page
